wip: MagicToast > refactor > demo
